Add route types
