+++ packages/contracts/source/contracts/p2p/server.ts
@@ -1,4 +1,5 @@
-import { Request, ResponseToolkit, ServerInjectOptions, ServerInjectResponse, ServerRoute } from "@hapi/hapi";
+import { Request, ResponseToolkit, Server as HapiServer, ServerInjectOptions, ServerInjectResponse, ServerRoute } from "@hapi/hapi";
+import { Schema } from "joi";
 
 export interface Controller {
 	handle(request: Request, h: ResponseToolkit): Promise<any>;
@@ -11,4 +12,28 @@ export interface Server {
 	register(plugins: any): Promise<void>; // TODO: Add proper types
 	route(routes: ServerRoute | ServerRoute[]): Promise<void>;
 	inject(options: string | ServerInjectOptions): Promise<ServerInjectResponse>;
+}
+
+export type Codec = {
+	request: {
+		serialize: any;
+		deserialize: any;
+	};
+	response: {
+		serialize: any;
+		deserialize: any;
+	};
+};
+
+export type RouteConfig = {
+	id: string;
+	validation?: Schema;
+	codec: Codec;
+	maxBytes?: number;
+};
+
+export interface Route {
+	register(server: HapiServer): void;
+	getRoutesConfigByPath(): { [path: string]: RouteConfig };
+
 }
