+++ packages/p2p/source/socket-server/plugins/codec.ts
@@ -2,7 +2,6 @@ import { inject, injectable, multiInject, tagged } from "@mainsail/container";
 import { Contracts, Identifiers } from "@mainsail/contracts";
 import { Providers } from "@mainsail/kernel";
 
-import { Route } from "../routes/route.js";
 import { BasePlugin } from "./base-plugin.js";
 
 @injectable()
@@ -18,7 +17,7 @@ export class CodecPlugin extends BasePlugin {
 	private readonly configuration!: Providers.PluginConfiguration;
 
 	@multiInject(Identifiers.P2P.Routes)
-	private readonly routes!: Route[];
+	private readonly routes!: Contracts.P2P.Route[];
 
 	public register(server) {
 		if (this.configuration.getRequired("developmentMode.enabled")) {
