+++ packages/p2p/source/socket-server/routes/get-peers.ts
@@ -1,14 +1,15 @@
 import { injectable } from "@mainsail/container";
+import { Contracts } from "@mainsail/contracts";
 
 import { Routes } from "../../enums.js";
 import { Codecs } from "../codecs/index.js";
 import { GetPeersController } from "../controllers/index.js";
 import { Schemas } from "../schemas/index.js";
-import { Route, RouteConfig } from "./route.js";
+import { Route } from "./route.js";
 
 @injectable()
 export class GetPeersRoute extends Route {
-	public getRoutesConfigByPath(): { [path: string]: RouteConfig } {
+	public getRoutesConfigByPath(): { [path: string]: Contracts.P2P.RouteConfig } {
 		return {
 			"/getPeers": {
 				codec: Codecs.getPeers,
