+++ packages/p2p/source/socket-server/routes/post-prevote.ts
@@ -1,15 +1,16 @@
 import { injectable } from "@mainsail/container";
+import { Contracts } from "@mainsail/contracts";
 
 import { constants } from "../../constants.js";
 import { Routes } from "../../enums.js";
 import { Codecs } from "../codecs/index.js";
 import { PostPrevoteController } from "../controllers/index.js";
 import { Schemas } from "../schemas/index.js";
-import { Route, RouteConfig } from "./route.js";
+import { Route } from "./route.js";
 
 @injectable()
 export class PostPrevoteRoute extends Route {
-	public getRoutesConfigByPath(): { [path: string]: RouteConfig } {
+	public getRoutesConfigByPath(): { [path: string]: Contracts.P2P.RouteConfig } {
 		return {
 			"/postPrevote": {
 				codec: Codecs.postPrevote,
