+++ packages/p2p/source/socket-server/server.ts
@@ -5,7 +5,6 @@ import { Contracts, Identifiers } from "@mainsail/contracts";
 import { constants } from "../constants.js";
 import { plugin as hapiNesPlugin } from "../hapi-nes/index.js";
 import { AcceptPeerPlugin, CodecPlugin, HeaderHandlePlugin, HeaderIncludePlugin, RateLimitPlugin, StatisticPlugin, ValidateDataPlugin, ValidateIpPlugin } from "./plugins/index.js";
-import { Route } from "./routes/route.js";
 
 @injectable()
 export class Server implements Contracts.P2P.Server {
@@ -16,7 +15,7 @@ export class Server implements Contracts.P2P.Server {
 	private readonly logger!: Contracts.Kernel.Logger;
 
 	@multiInject(Identifiers.P2P.Routes)
-	private readonly routes!: Route[];
+	private readonly routes!: Contracts.P2P.Route[];
 
 	private server!: HapiServer;
